+++ drake/systems/plants/RigidBody.cpp
@@ -18,7 +18,6 @@ RigidBody::RigidBody()
     : collision_filter_group(DrakeCollision::DEFAULT_GROUP),
       collision_filter_ignores(DrakeCollision::NONE_MASK),
       parent(nullptr) {
-  robotnum = 0;
   position_num_start = 0;
   velocity_num_start = 0;
   body_index = 0;
@@ -27,13 +26,17 @@ RigidBody::RigidBody()
   I << drake::SquareTwistMatrix<double>::Zero();
 }
 
-const std::string& RigidBody::name() const { return name_; }
+const std::string& RigidBody::get_name() const { return name_; }
 
-const std::string& RigidBody::model_name() const { return model_name_; }
+void RigidBody::set_name(const std::string& name) { name_ = name; }
 
-int RigidBody::get_model_id() const { return robotnum; }
+const std::string& RigidBody::get_model_name() const { return model_name_; }
 
-void RigidBody::set_model_id(int model_id) { robotnum = model_id; }
+void RigidBody::set_model_name(const std::string& name) { model_name_ = name; }
+
+int RigidBody::get_model_id() const { return model_id_; }
+
+void RigidBody::set_model_id(int model_id) { model_id_ = model_id; }
 
 void RigidBody::setJoint(std::unique_ptr<DrakeJoint> new_joint) {
   this->joint = move(new_joint);
@@ -121,7 +124,7 @@ RigidBody::CollisionElement::CollisionElement(
   // Collision elements should be set to static in a later Initialize() stage as
   // described in issue #2661.
   // TODO(amcastro-tri): remove this hack.
-  if (body->name() == "world") set_static();
+  if (body->get_name() == "world") set_static();
 }
 
 RigidBody::CollisionElement* RigidBody::CollisionElement::clone() const {