+++ drake/systems/plants/RigidBodySystem.cpp
@@ -307,15 +307,15 @@ DRAKERBSYSTEM_EXPORT RigidBodySystem::StateVector<double> getInitialState(
     std::list<RelativePositionConstraint> constraints;
     for (const auto& loop : loops) {
       constraints.push_back(RelativePositionConstraint(
-          sys.tree.get(), zero, zero, zero, loop.frameA->frame_index,
-          loop.frameB->frame_index, bTbp, tspan));
+          sys.tree.get(), zero, zero, zero, loop.frameA_->frame_index,
+          loop.frameB_->frame_index, bTbp, tspan));
       std::shared_ptr<SingleTimeKinematicConstraintWrapper> con1wrapper(
           new SingleTimeKinematicConstraintWrapper(&constraints.back(),
                                                    &kin_helper));
       prog.AddGenericConstraint(con1wrapper, {qvar});
       constraints.push_back(RelativePositionConstraint(
-          sys.tree.get(), loop.axis, loop.axis, loop.axis,
-          loop.frameA->frame_index, loop.frameB->frame_index, bTbp,
+          sys.tree.get(), loop.axis_, loop.axis_, loop.axis_,
+          loop.frameA_->frame_index, loop.frameB_->frame_index, bTbp,
           tspan));
       std::shared_ptr<SingleTimeKinematicConstraintWrapper> con2wrapper(
           new SingleTimeKinematicConstraintWrapper(&constraints.back(),
@@ -396,7 +396,7 @@ RigidBodySpringDamper::RigidBodySpringDamper(RigidBodySystem& sys,
 }
 
 const std::string& RigidBodySensor::get_model_name() const {
-  return frame_->body->model_name();
+  return frame_->body->get_model_name();
 }
 
 const RigidBodyFrame& RigidBodySensor::get_frame() const {
@@ -748,8 +748,8 @@ void parseSDFLink(RigidBodySystem& sys, int model_id, XMLElement* node,
   XMLElement* pose = node->FirstChildElement("pose");
   if (pose) {
     poseValueToTransform(pose, pose_map, transform_link_to_model);
-    pose_map.insert(
-        std::pair<string, Isometry3d>(body->name_, transform_link_to_model));
+    pose_map.insert(std::pair<string, Isometry3d>(body->get_name(),
+                                                  transform_link_to_model));
   }
 
   // Processes each sensor element within the link.
@@ -889,7 +889,7 @@ void RigidBodySystem::addRobotFromURDFString(
     const DrakeJoint::FloatingBaseType floating_base_type) {
   // first add the urdf to the rigid body tree
   drake::parsers::urdf::AddRobotFromURDFString(urdf_string, root_dir,
-    floating_base_type, tree.get());
+                                               floating_base_type, tree.get());
 
   // now parse additional tags understood by rigid body system (actuators,
   // sensors, etc)
@@ -908,7 +908,7 @@ void RigidBodySystem::addRobotFromURDF(
     std::shared_ptr<RigidBodyFrame> weld_to_frame) {
   // Adds the URDF to the rigid body tree.
   drake::parsers::urdf::AddRobotFromURDF(urdf_filename, floating_base_type,
-    weld_to_frame, tree.get());
+                                         weld_to_frame, tree.get());
 
   // Parses additional tags understood by rigid body system (e.g., actuators,
   // sensors, etc).
@@ -929,7 +929,7 @@ void RigidBodySystem::addRobotFromSDF(
     std::shared_ptr<RigidBodyFrame> weld_to_frame) {
   // Adds the robot to the rigid body tree.
   drake::parsers::sdf::AddRobotFromSDF(sdf_filename, floating_base_type,
-    weld_to_frame, tree.get());
+                                       weld_to_frame, tree.get());
 
   // Parses the additional SDF elements that are understood by RigidBodySystem,
   // namely (actuators, sensors, etc.).