+++ drake/systems/plants/RigidBodyTree.cpp
@@ -1,4 +1,4 @@
-#include "drake/systems/plants/RigidBodyTree.h"
+  #include "drake/systems/plants/RigidBodyTree.h"
 
 #include "drake/common/constants.h"
 #include "drake/common/eigen_types.h"
@@ -63,7 +63,7 @@ RigidBodyTree::RigidBodyTree(
     const DrakeJoint::FloatingBaseType floating_base_type)
     : RigidBodyTree() {
   drake::parsers::urdf::AddRobotFromURDF(urdf_filename, floating_base_type,
-    this);
+                                         this);
 }
 
 RigidBodyTree::RigidBodyTree(void)
@@ -73,8 +73,8 @@ RigidBodyTree::RigidBodyTree(void)
 
   // Adds the rigid body representing the world.
   std::unique_ptr<RigidBody> b(new RigidBody());
-  b->name_ = std::string(RigidBodyTree::kWorldLinkName);
-  b->robotnum = 0;
+  b->set_name(RigidBodyTree::kWorldLinkName);
+  b->set_model_id(0);
   b->body_index = 0;
   bodies.push_back(std::move(b));
 }
@@ -119,7 +119,7 @@ void RigidBodyTree::SortTree() {
 const RigidBodyActuator& RigidBodyTree::GetActuator(
     const std::string& name) const {
   for (const auto& actuator : actuators) {
-    if (actuator.name == name) {
+    if (actuator.name_ == name) {
       return actuator;
     }
   }
@@ -153,8 +153,8 @@ void RigidBodyTree::compile(void) {
       if (!hasChild) {
         // now check if this body is attached by a loop joint
         for (const auto& loop : loops) {
-          if ((loop.frameA->body == bodies[i].get()) ||
-              (loop.frameB->body == bodies[i].get())) {
+          if ((loop.frameA_->body == bodies[i].get()) ||
+              (loop.frameB_->body == bodies[i].get())) {
             hasChild = true;
             break;
           }
@@ -193,9 +193,9 @@ void RigidBodyTree::compile(void) {
   B.resize(num_velocities_, actuators.size());
   B = MatrixXd::Zero(num_velocities_, actuators.size());
   for (size_t ia = 0; ia < actuators.size(); ia++)
-    for (int i = 0; i < actuators[ia].body->getJoint().getNumVelocities(); i++)
-      B(actuators[ia].body->velocity_num_start + i, ia) =
-          actuators[ia].reduction;
+    for (int i = 0; i < actuators[ia].body_->getJoint().getNumVelocities(); i++)
+      B(actuators[ia].body_->velocity_num_start + i, ia) =
+          actuators[ia].reduction_;
 
   // Initializes the joint limit vectors.
   joint_limit_min = VectorXd::Constant(
@@ -290,14 +290,15 @@ void RigidBodyTree::drawKinematicTree(
   dotfile.open(graphviz_dotfile_filename);
   dotfile << "digraph {" << endl;
   for (const auto& body : bodies) {
-    dotfile << "  " << body->name_ << " [label=\"" << body->name_ << endl;
+    dotfile << "  " << body->get_name() << " [label=\"" << body->get_name()
+            << endl;
     dotfile << "mass=" << body->mass << ", com=" << body->com.transpose()
             << endl;
     dotfile << "inertia=" << endl << body->I << endl;
     dotfile << "\"];" << endl;
     if (body->hasParent()) {
       const auto& joint = body->getJoint();
-      dotfile << "  " << body->name_ << " -> " << body->parent->name_
+      dotfile << "  " << body->get_name() << " -> " << body->parent->get_name()
               << " [label=\"" << joint.getName() << endl;
       dotfile << "transform_to_parent_body=" << endl
               << joint.getTransformToParentBody().matrix() << endl;
@@ -308,7 +309,7 @@ void RigidBodyTree::drawKinematicTree(
   for (const auto& frame : frames) {
     dotfile << "  " << frame->name << " [label=\"" << frame->name
             << " (frame)\"];" << endl;
-    dotfile << "  " << frame->name << " -> " << frame->body->name_
+    dotfile << "  " << frame->name << " -> " << frame->body->get_name()
             << " [label=\"";
     dotfile << "transform_to_body=" << endl
             << frame->transform_to_body.matrix() << endl;
@@ -316,12 +317,12 @@ void RigidBodyTree::drawKinematicTree(
   }
 
   for (const auto& loop : loops) {
-    dotfile << "  " << loop.frameA->body->name_ << " -> "
-            << loop.frameB->body->name_ << " [label=\"loop " << endl;
+    dotfile << "  " << loop.frameA_->body->get_name() << " -> "
+            << loop.frameB_->body->get_name() << " [label=\"loop " << endl;
     dotfile << "transform_to_parent_body=" << endl
-            << loop.frameA->transform_to_body.matrix() << endl;
+            << loop.frameA_->transform_to_body.matrix() << endl;
     dotfile << "transform_to_child_body=" << endl
-            << loop.frameB->transform_to_body.matrix() << endl;
+            << loop.frameB_->transform_to_body.matrix() << endl;
     dotfile << "\"];" << endl;
   }
   dotfile << "}" << endl;
@@ -483,8 +484,8 @@ bool RigidBodyTree::collisionDetect(
   for (const int& body_idx : bodies_idx) {
     if (body_idx >= 0 && body_idx < static_cast<int>(bodies.size())) {
       for (const string& group : active_element_groups) {
-        bodies[body_idx]->appendCollisionElementIdsFromThisBody(
-            group, ids_to_check);
+        bodies[body_idx]->appendCollisionElementIdsFromThisBody(group,
+                                                                ids_to_check);
       }
     }
   }
@@ -560,10 +561,8 @@ void RigidBodyTree::potentialCollisions(const KinematicsCache<double>& cache,
   bodyB_idx.clear();
 
   for (size_t i = 0; i < num_potential_collisions; i++) {
-    const DrakeCollision::Element* elementA =
-        potential_collisions[i].elementA;
-    const DrakeCollision::Element* elementB =
-        potential_collisions[i].elementB;
+    const DrakeCollision::Element* elementA = potential_collisions[i].elementA;
+    const DrakeCollision::Element* elementB = potential_collisions[i].elementB;
     xA.col(i) = potential_collisions[i].ptA;
     xB.col(i) = potential_collisions[i].ptB;
     normal.col(i) = potential_collisions[i].normal;
@@ -578,27 +577,17 @@ RigidBodyTree::ComputeMaximumDepthCollisionPoints(
     const KinematicsCache<double>& cache, bool use_margins) {
   updateDynamicCollisionElements(cache);
   vector<DrakeCollision::PointPair> contact_points;
-  collision_model->ComputeMaximumDepthCollisionPoints(
-      use_margins, contact_points);
+  collision_model->ComputeMaximumDepthCollisionPoints(use_margins,
+                                                      contact_points);
   size_t num_contact_points = contact_points.size();
 
   for (size_t i = 0; i < num_contact_points; i++) {
     // Get bodies' transforms.
     const RigidBody& bodyA = *contact_points[i].elementA->get_body();
-    Isometry3d TA;
-    if (bodyA.hasParent()) {
-      TA = cache.getElement(bodyA).transform_to_world;
-    } else {  // body is the world.
-      TA = Isometry3d::Identity();
-    }
+    const Isometry3d& TA = cache.getElement(bodyA).transform_to_world;
 
     const RigidBody& bodyB = *contact_points[i].elementB->get_body();
-    Isometry3d TB;
-    if (bodyB.hasParent()) {
-      TB = cache.getElement(bodyB).transform_to_world;
-    } else {  // body is the world.
-      TB = Isometry3d::Identity();
-    }
+    const Isometry3d& TB = cache.getElement(bodyB).transform_to_world;
 
     // Transform to bodies' frames.
     // Note:
@@ -917,7 +906,7 @@ bool RigidBodyTree::isBodyPartOfRobot(const RigidBody& body,
     }
   }
 
-  return robotnum.find(body.robotnum) != robotnum.end();
+  return robotnum.find(body.get_model_id()) != robotnum.end();
 }
 
 double RigidBodyTree::getMass(const std::set<int>& robotnum) const {
@@ -1084,8 +1073,8 @@ KinematicPath RigidBodyTree::findKinematicPath(
 
   if (!least_common_ancestor_found) {
     std::ostringstream stream;
-    stream << "There is no path between " << bodies[start_body]->name_
-           << " and " << bodies[end_body]->name_ << ".";
+    stream << "There is no path between " << bodies[start_body]->get_name()
+           << " and " << bodies[end_body]->get_name() << ".";
     throw std::runtime_error(stream.str());
   }
   int least_common_ancestor = *start_body_lca_it;
@@ -1593,8 +1582,8 @@ RigidBodyTree::relativeQuaternionJacobianDotTimesV(
       static_cast<typename Gradient<decltype(Phi_autodiff), Dynamic>::type*>(
           nullptr));
   auto Phidot_vector = autoDiffToGradientMatrix(Phi_autodiff);
-  Map<Matrix<Scalar, kQuaternionSize, kSpaceDimension>>
-      Phid(Phidot_vector.data());
+  Map<Matrix<Scalar, kQuaternionSize, kSpaceDimension>> Phid(
+      Phidot_vector.data());
 
   const auto J_geometric_dot_times_v = geometricJacobianDotTimesV(
       cache, to_body_or_frame_ind, from_body_or_frame_ind, expressed_in);
@@ -1675,7 +1664,7 @@ RigidBody* RigidBodyTree::FindBody(const std::string& body_name,
     if (model_id != -1 && model_id != bodies[ii]->get_model_id()) continue;
 
     // Obtains a lower case version of the current body's model name.
-    string current_model_name = bodies[ii]->model_name();
+    string current_model_name = bodies[ii]->get_model_name();
     std::transform(current_model_name.begin(), current_model_name.end(),
                    current_model_name.begin(), ::tolower);
 
@@ -1685,7 +1674,7 @@ RigidBody* RigidBodyTree::FindBody(const std::string& body_name,
       continue;
 
     // Obtains a lower case version of the current body's name.
-    string current_body_name = bodies[ii]->name_;
+    string current_body_name = bodies[ii]->get_name();
     std::transform(current_body_name.begin(), current_body_name.end(),
                    current_body_name.begin(), ::tolower);
 
@@ -1814,7 +1803,7 @@ RigidBody* RigidBodyTree::findJoint(const std::string& joint_name,
   if (model_id != -1) {
     for (size_t ii = 0; ii < this->bodies.size(); ii++) {
       if (name_match[ii]) {
-        name_match[ii] = this->bodies[ii]->robotnum == model_id;
+        name_match[ii] = this->bodies[ii]->get_model_id() == model_id;
       }
     }
   }
@@ -1853,7 +1842,7 @@ int RigidBodyTree::findJointId(const std::string& joint_name, int robot) const {
 
 std::string RigidBodyTree::getBodyOrFrameName(int body_or_frame_id) const {
   if (body_or_frame_id >= 0) {
-    return bodies[body_or_frame_id]->name_;
+    return bodies[body_or_frame_id]->get_name();
   } else if (body_or_frame_id < -1) {
     return frames[-body_or_frame_id - 2]->name;
   } else {
@@ -1868,15 +1857,16 @@ Matrix<Scalar, Eigen::Dynamic, 1> RigidBodyTree::positionConstraints(
   for (size_t i = 0; i < loops.size(); i++) {
     {  // position constraint
       auto ptA_in_B =
-          transformPoints(cache, Vector3d::Zero(), loops[i].frameA->frame_index,
-                          loops[i].frameB->frame_index);
+          transformPoints(cache, Vector3d::Zero(),
+                          loops[i].frameA_->frame_index,
+                          loops[i].frameB_->frame_index);
       ret.template middleRows<3>(6 * i) = ptA_in_B;
     }
     {  // second position constraint (to constrain orientation)
       auto axis_A_end_in_B =
-          transformPoints(cache, loops[i].axis, loops[i].frameA->frame_index,
-                          loops[i].frameB->frame_index);
-      ret.template middleRows<3>(6 * i + 3) = axis_A_end_in_B - loops[i].axis;
+          transformPoints(cache, loops[i].axis_, loops[i].frameA_->frame_index,
+                          loops[i].frameB_->frame_index);
+      ret.template middleRows<3>(6 * i + 3) = axis_A_end_in_B - loops[i].axis_;
     }
   }
   return ret;
@@ -1892,12 +1882,12 @@ RigidBodyTree::positionConstraintsJacobian(const KinematicsCache<Scalar>& cache,
   for (size_t i = 0; i < loops.size(); i++) {
     // position constraint
     ret.template middleRows<3>(6 * i) = transformPointsJacobian(
-        cache, Vector3d::Zero(), loops[i].frameA->frame_index,
-        loops[i].frameB->frame_index, in_terms_of_qdot);
+        cache, Vector3d::Zero(), loops[i].frameA_->frame_index,
+        loops[i].frameB_->frame_index, in_terms_of_qdot);
     // second position constraint (to constrain orientation)
     ret.template middleRows<3>(6 * i + 3) = transformPointsJacobian(
-        cache, loops[i].axis, loops[i].frameA->frame_index,
-        loops[i].frameB->frame_index, in_terms_of_qdot);
+        cache, loops[i].axis_, loops[i].frameA_->frame_index,
+        loops[i].frameB_->frame_index, in_terms_of_qdot);
   }
   return ret;
 }
@@ -1911,12 +1901,12 @@ RigidBodyTree::positionConstraintsJacDotTimesV(
   for (size_t i = 0; i < loops.size(); i++) {
     // position constraint
     ret.template middleRows<3>(6 * i) = transformPointsJacobianDotTimesV(
-        cache, Vector3d::Zero(), loops[i].frameA->frame_index,
-        loops[i].frameB->frame_index);
+        cache, Vector3d::Zero(), loops[i].frameA_->frame_index,
+        loops[i].frameB_->frame_index);
     // second position constraint (to constrain orientation)
     ret.template middleRows<3>(6 * i + 3) = transformPointsJacobianDotTimesV(
-        cache, loops[i].axis, loops[i].frameA->frame_index,
-        loops[i].frameB->frame_index);
+        cache, loops[i].axis_, loops[i].frameA_->frame_index,
+        loops[i].frameB_->frame_index);
   }
   return ret;
 }
@@ -2028,8 +2018,8 @@ int RigidBodyTree::AddFloatingJoint(
 
       Eigen::Isometry3d transform_to_model = Eigen::Isometry3d::Identity();
       if (pose_map != nullptr &&
-          pose_map->find(bodies[i]->name_) != pose_map->end())
-        transform_to_model = pose_map->at(bodies[i]->name_);
+          pose_map->find(bodies[i]->get_name()) != pose_map->end())
+        transform_to_model = pose_map->at(bodies[i]->get_name());
 
       switch (floating_base_type) {
         case DrakeJoint::FIXED: {
@@ -2076,7 +2066,8 @@ void RigidBodyTree::addRobotFromURDFString(
     std::shared_ptr<RigidBodyFrame> weld_to_frame) {
   PackageMap package_map;
   drake::parsers::urdf::AddRobotFromURDFString(xml_string, package_map,
-    root_dir, floating_base_type, weld_to_frame, this);
+                                               root_dir, floating_base_type,
+                                               weld_to_frame, this);
 }
 
 // TODO(liang.fok) Remove this deprecated method prior to release 1.0.
@@ -2087,7 +2078,8 @@ void RigidBodyTree::addRobotFromURDFString(
     const DrakeJoint::FloatingBaseType floating_base_type,
     std::shared_ptr<RigidBodyFrame> weld_to_frame) {
   drake::parsers::urdf::AddRobotFromURDFString(xml_string, package_map,
-    root_dir, floating_base_type, weld_to_frame, this);
+                                               root_dir, floating_base_type,
+                                               weld_to_frame, this);
 }
 
 // TODO(liang.fok) Remove this deprecated method prior to release 1.0.
@@ -2096,8 +2088,8 @@ void RigidBodyTree::addRobotFromURDF(
     const DrakeJoint::FloatingBaseType floating_base_type,
     std::shared_ptr<RigidBodyFrame> weld_to_frame) {
   PackageMap package_map;
-  drake::parsers::urdf::AddRobotFromURDF(urdf_filename, package_map,
-    floating_base_type, weld_to_frame, this);
+  drake::parsers::urdf::AddRobotFromURDF(
+      urdf_filename, package_map, floating_base_type, weld_to_frame, this);
 }
 
 // TODO(liang.fok) Remove this deprecated method prior to release 1.0.
@@ -2106,8 +2098,8 @@ void RigidBodyTree::addRobotFromURDF(
     std::map<std::string, std::string>& package_map,
     const DrakeJoint::FloatingBaseType floating_base_type,
     std::shared_ptr<RigidBodyFrame> weld_to_frame) {
-  drake::parsers::urdf::AddRobotFromURDF(urdf_filename, package_map,
-    floating_base_type, weld_to_frame, this);
+  drake::parsers::urdf::AddRobotFromURDF(
+      urdf_filename, package_map, floating_base_type, weld_to_frame, this);
 }
 
 // TODO(liang.fok) Remove this deprecated method prior to release 1.0.
@@ -2116,7 +2108,7 @@ void RigidBodyTree::addRobotFromSDF(
     const DrakeJoint::FloatingBaseType floating_base_type,
     std::shared_ptr<RigidBodyFrame> weld_to_frame) {
   drake::parsers::sdf::AddRobotFromSDF(sdf_filename, floating_base_type,
-                                         weld_to_frame, this);
+                                       weld_to_frame, this);
 }
 
 // Explicit template instantiations for massMatrix.
@@ -2452,7 +2444,7 @@ template DRAKERBM_EXPORT KinematicsCache<double> RigidBodyTree::doKinematics(
 template DRAKERBM_EXPORT KinematicsCache<AutoDiffXd>
 RigidBodyTree::doKinematics(
     Eigen::MatrixBase<VectorX<AutoDiffXd>> const&) const;
-template DRAKERBM_EXPORT KinematicsCache<double>RigidBodyTree::doKinematics(
+template DRAKERBM_EXPORT KinematicsCache<double> RigidBodyTree::doKinematics(
     Eigen::MatrixBase<Eigen::Block<VectorXd, -1, 1, false>> const&) const;
 
 // Explicit template instantiations for doKinematics(q, v).