+++ drake/systems/plants/test/testIKtraj.cpp
@@ -4,10 +4,10 @@
 #include "gtest/gtest.h"
 
 #include "drake/Path.h"
+#include "drake/systems/plants/constraint/RigidBodyConstraint.h"
 #include "drake/systems/plants/IKoptions.h"
 #include "drake/systems/plants/RigidBodyIK.h"
 #include "drake/systems/plants/RigidBodyTree.h"
-#include "../constraint/RigidBodyConstraint.h"
 
 using Eigen::MatrixXd;
 using Eigen::Vector2d;
@@ -22,7 +22,7 @@ GTEST_TEST(testIKtraj, testIKtraj) {
 
   int r_hand{};
   for (int i = 0; i < static_cast<int>(model.bodies.size()); i++) {
-    if (model.bodies[i]->name_.compare(std::string("r_hand"))) {
+    if (model.bodies[i]->get_name().compare(std::string("r_hand"))) {
       r_hand = i;
     }
   }
@@ -70,26 +70,23 @@ GTEST_TEST(testIKtraj, testIKtraj) {
   MatrixXd qddot_sol(model.number_of_positions(), nT);
   int info = 0;
   std::vector<std::string> infeasible_constraint;
-  inverseKinTraj(&model, nT, t.data(), qdot0, q0, q0,
-                 constraint_array.size(), constraint_array.data(),
-                 ikoptions,
-                 &q_sol, &qdot_sol, &qddot_sol, &info, &infeasible_constraint);
+  inverseKinTraj(&model, nT, t.data(), qdot0, q0, q0, constraint_array.size(),
+                 constraint_array.data(), ikoptions, &q_sol, &qdot_sol,
+                 &qddot_sol, &info, &infeasible_constraint);
   EXPECT_EQ(info, 1);
 
   ikoptions.setFixInitialState(false);
   ikoptions.setMajorIterationsLimit(500);
-  inverseKinTraj(&model, nT, t.data(), qdot0, q0, q0,
-                 constraint_array.size(), constraint_array.data(),
-                 ikoptions,
-                 &q_sol, &qdot_sol, &qddot_sol, &info, &infeasible_constraint);
+  inverseKinTraj(&model, nT, t.data(), qdot0, q0, q0, constraint_array.size(),
+                 constraint_array.data(), ikoptions, &q_sol, &qdot_sol,
+                 &qddot_sol, &info, &infeasible_constraint);
   EXPECT_EQ(info, 1);
 
   Eigen::RowVectorXd t_inbetween(5);
   t_inbetween << 0.1, 0.15, 0.3, 0.4, 0.6;
   ikoptions.setAdditionaltSamples(t_inbetween);
-  inverseKinTraj(&model, nT, t.data(), qdot0, q0, q0,
-                 constraint_array.size(), constraint_array.data(),
-                 ikoptions,
-                 &q_sol, &qdot_sol, &qddot_sol, &info, &infeasible_constraint);
+  inverseKinTraj(&model, nT, t.data(), qdot0, q0, q0, constraint_array.size(),
+                 constraint_array.data(), ikoptions, &q_sol, &qdot_sol,
+                 &qddot_sol, &info, &infeasible_constraint);
   EXPECT_EQ(info, 1);
 }